Update Dockerfile version and revert changes to index.ts
- Bumped version label in Dockerfile from 0.1.0 to 1.0.0.
- Reverted index.ts and server.ts to modular structure
